# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Patient Tracker App",
        "contact": {
            "name": "Mahmoud Mohamed",
            "url": "https://www.linkedin.com/in/mahmoud-mohamed-955932b5/",
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "rep1@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "reset password",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "rep1@gmail.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout currently logged in user",
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "refreshes expired token",
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}